backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-groq
pypdf
sentence-transformers
